+++ docs/changes.rst
@@ -6,6 +6,7 @@ Not yet released.
 **New features**
 
 * :doc:`/formats/markdown` support, thanks to Anders Kaplan.
+* :ref:`category` can now organize components within a project.
 * :doc:`/formats/fluent` now has better syntax checks thanks to Henry Wilkes.
 * Inviting users now works with all authentication methods.
 * Docker container supports file backed secrets, see :ref:`docker-secrets`.